text_extraction_system/requirements.txt
JSON-log-formatter==0.2.0
Jinja2==2.11.3
MarkupSafe==2.0.1
Pillow==8.3.2
PyWavelets==1.3.0
aiofiles==0.6.0
amqp==5.0.1
backports.zoneinfo==0.2.1
billiard==3.6.3.0
camelot-py==0.8.2
celery==5.0.1
certifi==2022.6.15
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.0.12
click-didyoumean==0.3.0
click-repl==0.2.0
click==7.1.2
coverage==6.4.1
cryptography==37.0.4
dataclasses-json==0.5.2
dataclasses==0.6
dateparser==1.1.1
deskew==0.10.16
et-xmlfile==1.1.0
fastapi==0.61.1
fasttext==0.9.2
gensim==4.1.2
h11==0.13.0
https://github.com/LexPredict/pdfminer.six/archive/20210513-grid-fix.zip
idna==3.3
iniconfig==1.1.1
joblib==1.1.0
kombu==5.0.2
lxml==4.9.0
marshmallow-enum==1.5.1
marshmallow==3.8.0
msgpack==1.0.2
mypy-extensions==0.4.3
nltk==3.7
numpy==1.22.3
opencv-python==4.4.0.46
openpyxl==3.0.10
packaging==21.3
pandas==1.4.2
pdf2image==1.16.0
pikepdf==6.2.5
pip==20.0.2
pipdeptree==2.2.1
pluggy==0.13.1
prompt-toolkit==3.0.30
psutil==5.9.1
py==1.11.0
pycountry==22.3.5
pycparser==2.21
pydantic==1.8.2
pyparsing==3.0.9
pytest-asyncio==0.14.0
pytest==6.1.1
python-dateutil==2.8.2
python-dotenv==0.14.0
python-magic==0.4.26
python-multipart==0.0.5
redis==3.5.3
regex==2022.3.2
requests==2.27.1
scikit-image==0.19.3
scikit-learn==0.23.1
scipy==1.8.1
six==1.16.0
smart-open==6.0.0
sortedcontainers==2.4.0
starlette==0.13.6
stringcase==1.2.0
threadpoolctl==3.1.0
toml==0.10.2
tqdm==4.64.0
typing-extensions==4.3.0
uvicorn==0.12.1
vine==5.0.0
wcwidth==0.2.5
webdavclient3==3.14.5